Finished mongo unit tests. Renamed tests directory to test
