+++ package.json
@@ -6,8 +6,8 @@
   "scripts": {
     "start": "node server.js",
     "dev": "nodemon server.js",
-    "test": "nyc mocha --recursive ./tests/test-defaults.js ./tests/unit ./tests/integration",
-    "unit-test": "mocha --recursive ./tests/test-defaults.js ./tests/unit"
+    "test": "nyc mocha --recursive ./test/test-defaults.js ./test/unit ./test/integration",
+    "unit-test": "mocha --recursive ./test/test-defaults.js ./test/unit"
   },
   "engines": {
     "npm": "4.0.5",
