+++ test/integration/installApp.js
