+++ test/integration/multiVoting.js
